Fix: board entity에서 가져올 user 정보 수정

+++ taveboard/src/main/java/com/practice/taveboard/board/repository/BoardRepository.java
@@ -13,7 +13,7 @@ import java.util.Optional;
 @Repository
 public interface BoardRepository extends JpaRepository<Board, Long> {
 
-    Optional<Board> findById(int id);
+    Optional<Board> findById(Long id);
 
     @Query("select f from BOARD_TABLE f order by f.createdTime")
     Page<Board> findAllByCreatedTimeDesc(Pageable pageable);
